app/function_code/store-backend-ops-v2/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
azure-identity
azure-storage-blob
azure-servicebus
azure-cosmos

azure-monitor-opentelemetry
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry-exporter


asyncio
aiohttp

# Mostly Unused
azure-storage-queue
azure-appconfiguration
azure-eventhub
azure-eventhub-checkpointstoreblob
azure-eventhub-checkpointstoreblob-aio